third_party/blink/web_tests/storage/indexeddb/mozilla/key-requirements-delete-null-key-expected.txt

Test IndexedDB's behavior deleting entry with no key

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "key-requirements-delete-null-key.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('bar');
Expecting exception from objectStore.delete(null);
PASS Exception was thrown.
PASS code is 0
PASS ename is 'DataError'
Exception message: Failed to execute 'delete' on 'IDBObjectStore': No key or key range specified.
PASS successfullyParsed is true

TEST COMPLETE
